Revert "Signal.c and signal.h revision"
This reverts commit 0174ed6921f0747e22b675dcc1a3a4046fb3dffa.

+++ src/signal.c
@@ -9,66 +9,84 @@
 #include <fcntl.h>
 #include "xmod.h"
 
-int signSIGINTAnwserHandler(char *answer) {
-    if (answer == NULL) return -1;
+/**
+ * 1º-> funções de reigisto de sinais
+ * 2ª-> aplicar os signhandlers no xmod.c
+ * 3ª-> aplicar o sinal de fim de fil de processo filho
+ * 
+ */
 
-    int status;
-
-    switch (answer[0]) {
-    case '1':
-        for(int i = 0; i < numberOfChildren; i++) {
+int signSIGINTAnwserHandler(char answer[]){
+    //Se for igual a 1, vai mandar todos os procesos abortarem
+    if(answer[0] =='1'){
+        for(int i = 0; i<numberOfChildren; i++){
             kill(childProcesses[i], SIGUSR1);
             eventSignalSent(SIGUSR1, childProcesses[i]);
         }
+        //Espera que os filhos todos morram antes de morrer
+        for(int i = 0; i<numberOfChildren; i++){
+            int status;
+            waitpid(childProcesses[i], &status,1 );
+        }
 
-        // Waits for the children to terminate before it terminates itself
-        for(int i = 0; i<numberOfChildren; i++)
-            waitpid(childProcesses[i], &status, 1);
-
+        endLogFile();
         write(STDOUT_FILENO, "The program has been cancelled!\n", 33);
 
-        // Raises kill and registers it
-        eventSignalSent(SIGKILL, getpid());
-        eventSignalRecv(SIGKILL);
-        eventProcExit(0);
-        raise(SIGKILL);
-        
-    case '2':
-        for(int i = 0; i < numberOfChildren; i++) {
-            kill(childProcesses[i], SIGCONT);
+        //Levanta o kill e regista tudo
+        eventSignalSent(SIGKILL, getpid());   //Envio que vou me matar
+        eventSignalRecv(SIGKILL);     //Recebo que me vou matar
+        eventProcExit(0);    //Aviso que me matei
+        raise(SIGKILL);     //Mato-me
+        return 0; 
+
+    }
+
+    else if(answer[0] == '2'){
+        for(int i = 0; i<numberOfChildren; i++){
             eventSignalSent(SIGCONT, childProcesses[i]);
+            kill(childProcesses[i], SIGCONT);
         }
-        
+
         write(STDOUT_FILENO, "The program has been continued!\n", 33);
 
         return 0;
+
     }
-    return -1;
+
+    else{return -1;}
 }
 
-void signSIGINTHandler(int signo) {
-    // Registers the signal
-    eventSignalRecv(SIGINT);
 
-    // Writing the pid; directory_name; number_of_files_found; number_of_files_modified
+
+void signSIGINTHandler(int signo){
+    eventSignalRecv(SIGINT);
+    //Escrever a mensagem
     char buffer[50];
+
     sprintf(buffer, "%d; %s; %d; %d\n", getpid(), filename, nftot, nfmod);
     write(STDOUT_FILENO, buffer, strlen(buffer));
+    
 
+    //Se for pai:
     if (isFirstParent) {
-        sleep(5);  // Waiting for the children to write the messages
-
+        //Maneira de esperar que os filhos escrevam tudo
+        sleep(5);
         char input[1];
-        write(STDOUT_FILENO,"Do you want to stop the program(1) or do you prefer to continue(2)?\t", 68);
-        read(STDIN_FILENO, input, 1);
 
-        while (signSIGINTAnwserHandler(input) == -1) {
+
+        write(STDOUT_FILENO,"Do you want to stop the program(1) or do you prefer to continue(2)?\t", 69);
+    mauInput:
+        read(STDIN_FILENO, input, 1);
+        if(signSIGINTAnwserHandler(input)==-1){
             write(STDOUT_FILENO,"Please insert '1' to stop or '2' to continue!\t", 47);
-            read(STDIN_FILENO, input, 1);
+            goto mauInput;
         }
+
+        
     }
-    else {
-        // Children pause and registers the event
+    else
+    {
+        //senão entra em pausa
         eventSignalSent(SIGSTOP, getpid());
         eventSignalRecv(SIGSTOP);
         pause();
@@ -76,32 +94,37 @@ void signSIGINTHandler(int signo) {
 }
 
 
-void sighandlerSIGCONT(int signo) {
-    // Registers the signal
+void sighandlerSIGCONT(int signo){
+    //Regist que recebeu o sinal
     eventSignalRecv(SIGCONT);
 
-    // Sends the signals to the children and registers it
     for (int i = 0; i < numberOfChildren; i++) {
+        //envia o sigcont aos filhos e regista
         kill(childProcesses[i], SIGCONT);
         eventSignalSent(SIGCONT, childProcesses[i]);
     }
+
 }
 
-void sigHandlerSIGUSR1(int signo) {
-    // Registers the signal
+void sigHandlerSIGUSR1(int signo){
     eventSignalRecv(SIGUSR1);
 
-    // Sends the signals to the children and registers it
-    for (int i = 0; i < numberOfChildren; i++) {
+    for(int i = 0; i<numberOfChildren; i++){
         kill(childProcesses[i], SIGUSR1);
         eventSignalSent(SIGUSR1, childProcesses[i]);
     }
 
     eventSignalSent(SIGKILL, getpid());
     eventSignalRecv(SIGKILL);
-    eventSignalSent(SIGCHLD, getppid());
+    eventSignalSent(SIGCHLD, getppid());    //Aviso o paizinho que me vou matar
     eventProcExit(0);
     raise(SIGKILL);
+    
+    
+    
 }
 
-void sigHandlerSIGCHILD(int signo) {eventSignalRecv(SIGCHLD);}
+void sigHandlerSIGCHILD(int signo){
+    eventSignalRecv(SIGCHLD);
+}
+
